inference/frontend/requirements.txt
streamlit==1.31.0
requests==2.31.0
Pillow==10.2.0
